@@ -3,7 +3,7 @@ name = "entomological-label-information-extraction"
 version = "1.0.0"
 description = "AI-powered framework for semi-automated processing of entomological specimen labels"
 readme = "README.md"
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 license = { text = "MIT" }
 keywords = ["entomology", "machine-learning", "ocr", "computer-vision", "tensorflow"]
 classifiers = [
@@ -12,9 +12,9 @@ classifiers = [
     "License :: OSI Approved :: MIT License",
     "Operating System :: OS Independent",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
     "Topic :: Scientific/Engineering :: Artificial Intelligence",
     "Topic :: Scientific/Engineering :: Image Processing"
 ]
@@ -80,7 +80,7 @@ where = ["."]
 
 [tool.black]
 line-length = 88
-target-version = ['py39']
+target-version = ['py310']
 include = '\.pyi?$'
 extend-exclude = '''
 (
