@@ -9,9 +9,15 @@ import os
 import warnings
 
 # Suppress warning messages during execution
-warnings.filterwarnings('ignore')
+warnings.filterwarnings("ignore")
 
-def metrics(target: list, pred: pd.DataFrame, gt: pd.DataFrame, out_dir: Path = Path(os.getcwd())) -> str:
+
+def metrics(
+    target: list,
+    pred: pd.DataFrame,
+    gt: pd.DataFrame,
+    out_dir: Path = Path(os.getcwd()),
+) -> str:
     """
     Build a text report showing the main classification metrics,
     to measure the quality of predictions of the classification model, and save it to a text file.
@@ -29,21 +35,29 @@ def metrics(target: list, pred: pd.DataFrame, gt: pd.DataFrame, out_dir: Path = 
         report_file = out_dir / "classification_report.txt"
         accuracy = accuracy_score(gt, pred) * 100
         report = classification_report(gt, pred, target_names=target)
-        
+
         print("Accuracy Score ->", accuracy)
         print(report)
-        
-        with open(report_file, 'w') as file:
+
+        with open(report_file, "w") as file:
             file.write(f"Accuracy Score -> {accuracy}\n")
             file.write(report)
-        
-        print(f"\nThe Classification Report has been successfully saved in {report_file}")
+
+        print(
+            f"\nThe Classification Report has been successfully saved in {report_file}"
+        )
         return report
     except Exception as e:
         print(f"Error computing classification metrics: {e}")
         return ""
 
-def cm(target: list, pred: pd.DataFrame, gt: pd.DataFrame, out_dir: Path = Path(os.getcwd())) -> None:
+
+def cm(
+    target: list,
+    pred: pd.DataFrame,
+    gt: pd.DataFrame,
+    out_dir: Path = Path(os.getcwd()),
+) -> None:
     """
     Compute confusion matrix to evaluate the performance of the classification.
 
@@ -55,21 +69,28 @@ def cm(target: list, pred: pd.DataFrame, gt: pd.DataFrame, out_dir: Path = Path(
     """
     try:
         cm = confusion_matrix(gt, pred)
-        cmn = cm.astype('float') / cm.sum(axis=1)[:, np.newaxis]
-        fig, ax = plt.subplots(figsize=(15,10))
-        matrix = sns.heatmap(cmn, annot=True, fmt='.2f', xticklabels=target, yticklabels=target, cmap="Greens",
-                              annot_kws={"size": 14})
-        plt.ylabel('Ground truth', fontsize=18)
-        plt.xlabel('Predictions', labelpad=30, fontsize=18)
+        cmn = cm.astype("float") / cm.sum(axis=1)[:, np.newaxis]
+        fig, ax = plt.subplots(figsize=(15, 10))
+        matrix = sns.heatmap(
+            cmn,
+            annot=True,
+            fmt=".2f",
+            xticklabels=target,
+            yticklabels=target,
+            cmap="Greens",
+            annot_kws={"size": 14},
+        )
+        plt.ylabel("Ground truth", fontsize=18)
+        plt.xlabel("Predictions", labelpad=30, fontsize=18)
         plt.xticks(fontsize=16)
         plt.yticks(fontsize=16)
-        
+
         figure = matrix.get_figure()
         filename = f"{out_dir.stem}_cm.png"
         cm_path = out_dir / filename
         figure.savefig(cm_path)
         plt.close(fig)
-        
+
         print(f"\nThe Confusion Matrix has been successfully saved in {cm_path}")
     except Exception as e:
         print(f"Error generating confusion matrix: {e}")
