@@ -2,16 +2,18 @@
 import json
 import nltk
 from nltk import word_tokenize
-nltk.download('punkt', quiet=True)
+
+nltk.download("punkt", quiet=True)
 import string
 import pandas as pd
 import re
 from typing import List, Dict
 
 # Constants
-NON_ASCII = re.compile(" [^\x00-\x7F] ")
-NON_ALPHA_NUM = re.compile("[^a-zA-Z\d\s]{2,}")
+NON_ASCII = re.compile(" [^\x00-\x7f] ")
+NON_ALPHA_NUM = re.compile(r"[^a-zA-Z\d\s]{2,}")
 PIPE = re.compile("[|]")
+
 
 def count_mean_token_length(tokens: List[str]) -> float:
     """
@@ -28,6 +30,7 @@ def count_mean_token_length(tokens: List[str]) -> float:
     total_length = sum(len(token) for token in tokens)
     mean_length = total_length / len(tokens)
     return round(mean_length, 2)
+
 
 def is_plausible_prediction(transcript: str) -> bool:
     """
@@ -47,35 +50,37 @@ def is_plausible_prediction(transcript: str) -> bool:
         print(f"Error checking plausible prediction: {e}")
         return False
 
+
 def correct_transcript(transcript: str) -> str:
     """
     Performs corrections on a transcript, removing non-ASCII characters, multiple non-alphanumeric characters,
     the pipe character, and other special symbols (like °, ', , etc.). Also removes any trailing periods.
-    
+
     Args:
         transcript (str): Input transcript.
-    
+
     Returns:
         str: Corrected transcript.
     """
     try:
         # Remove non-ASCII characters
-        transcript = re.sub(NON_ASCII, ' ', transcript)
-        
+        transcript = re.sub(NON_ASCII, " ", transcript)
+
         # Remove non-alphanumeric characters (like special symbols, except for spaces)
-        transcript = re.sub(NON_ALPHA_NUM, '', transcript)
-        
+        transcript = re.sub(NON_ALPHA_NUM, "", transcript)
+
         # Remove the pipe character
-        transcript = re.sub(PIPE, '', transcript)
-        
+        transcript = re.sub(PIPE, "", transcript)
+
         # Remove specific characters (degree symbol, apostrophes, commas)
-        transcript = transcript.replace('°', '').replace("'", '').replace(",", "")
-        
+        transcript = transcript.replace("°", "").replace("'", "").replace(",", "")
+
         # Remove any trailing periods
         return transcript.rstrip(".")
     except Exception as e:
         print(f"Error correcting transcript: {e}")
         return transcript
+
 
 def is_nuri(transcript: str) -> bool:
     """
@@ -89,6 +94,7 @@ def is_nuri(transcript: str) -> bool:
     """
     return transcript.startswith("http")
 
+
 def is_empty(transcript: str) -> bool:
     """
     Checks if a transcript is empty.
@@ -100,6 +106,7 @@ def is_empty(transcript: str) -> bool:
         bool: True if the transcript is empty, False otherwise.
     """
     return len(transcript.strip()) == 0
+
 
 def save_transcripts(transcripts: Dict, file_name: str) -> None:
     """
@@ -113,6 +120,7 @@ def save_transcripts(transcripts: Dict, file_name: str) -> None:
         pd.DataFrame.from_dict(transcripts, orient="index").to_csv(file_name)
     except Exception as e:
         print(f"Error saving transcripts to CSV: {e}")
+
 
 def save_json(transcripts: List[Dict], file_name: str) -> None:
     """
@@ -128,6 +136,7 @@ def save_json(transcripts: List[Dict], file_name: str) -> None:
     except Exception as e:
         print(f"Error saving JSON file: {e}")
 
+
 def process_ocr_output(ocr_output: str) -> None:
     """
     Processes OCR output, categorizing and saving transcripts based on Nuri, empty, plausible, and corrected.
@@ -137,14 +146,14 @@ def process_ocr_output(ocr_output: str) -> None:
     """
     try:
         nuri_labels, empty_labels, plausible_labels, clean_labels = {}, {}, [], []
-        
-        with open(ocr_output, 'r') as f:
+
+        with open(ocr_output, "r") as f:
             labels = json.load(f)
-            
+
         for label in labels:
             text = label.get("text", "")
             label_id = label.get("ID", "")
-            
+
             if is_nuri(text):
                 nuri_labels[label_id] = text
             elif is_empty(text):
@@ -152,7 +161,7 @@ def process_ocr_output(ocr_output: str) -> None:
             elif is_plausible_prediction(text):
                 plausible_labels.append({"ID": label_id, "text": text})
                 clean_labels.append({"ID": label_id, "text": correct_transcript(text)})
-        
+
         save_transcripts(nuri_labels, "nuris.csv")
         save_transcripts(empty_labels, "empty_transcripts.csv")
         save_json(plausible_labels, "plausible_transcripts.json")
